+++ image_client/picturae_csv_create.py
@@ -5,6 +5,9 @@
    Source for taxon names at IPNI (International Plant Names Index): https://www.ipni.org/ """
 from uuid import uuid4
 import traceback
+
+import pandas as pd
+
 import picturae_config
 from rpy2 import robjects
 from rpy2.robjects import pandas2ri
@@ -288,6 +291,31 @@ class CsvCreatePicturae(Importer):
 
         cursor.close()
 
+
+
+    def taxon_unmatch_create(self, unmatched_taxa: pd.DataFrame):
+        """taxon_unmatch_create: creates sql query for creating new records in taxa unmatch,
+                                from rows that did not pass TNRS successfully,
+                                either through spelling, or taxonomic errors.
+            args:
+                unmatched_taxa: a pandas dataframe with unmatched taxa terms filtered by score
+        """
+        print("uploading unmatched taxa")
+        for index, row in unmatched_taxa.iterrows():
+            catalognumber = unmatched_taxa.columns.get_loc("CatalogNumber")
+
+            sql = create_unmatch_tab(row=row, df=unmatched_taxa, tab_name='taxa_unmatch')
+
+            barcode_check = f'''SELECT CatalogNumber FROM taxa_unmatch 
+                                WHERE CatalogNumber = {row[catalognumber]}'''
+
+            # checking if unmatched taxa's barcode already on unmatch_taxa table
+            sql_result = self.specify_db_connection.get_one_record(barcode_check)
+            if sql_result is None:
+                self.create_table_record(sql)
+            else:
+                pass
+
     def taxon_check_real(self):
         """taxon_check_real:
            Sends the concatenated taxon column, through TNRS, to match names,
@@ -316,22 +344,9 @@ class CsvCreatePicturae(Importer):
         unmatched_taxa = resolved_taxon[resolved_taxon["overall_score"] < .99]
 
         # writing unmatched taxa to db table taxa_unmatch
+
         if len(unmatched_taxa) > 0:
-            print("uploading unmatched taxa")
-            for index, row in unmatched_taxa.iterrows():
-                catalognumber = unmatched_taxa.columns.get_loc("CatalogNumber")
-
-                sql = create_unmatch_tab(row=row, df=unmatched_taxa, tab_name='taxa_unmatch')
-
-                barcode_check = f'''SELECT CatalogNumber FROM taxa_unmatch 
-                                    WHERE CatalogNumber = {row[catalognumber]}'''
-
-                # checking if unmatched taxa's barcode already on unmatch_taxa table
-                sql_result = self.specify_db_connection.get_one_record(barcode_check)
-                if sql_result is None:
-                    self.create_table_record(sql)
-                else:
-                    pass
+            self.taxon_unmatch_create(unmatched_taxa)
 
         # filtering out taxa with tnrs scores lower than .99 (basically exact match)
         resolved_taxon = resolved_taxon[resolved_taxon["overall_score"] >= .99]
