+++ image_client/test_csv_purge_sql/picturae_purge_script.sql
@@ -1,13 +1,13 @@
 # removes all records added in the selected time window(in this case last 20 days), order matters
 # as certain tables external keys
-DELETE FROM casbotany.collectionobjectattachment WHERE TimestampCreated  > (now() - interval 20 day);
-DELETE FROM casbotany.attachment WHERE TimestampCreated  > (now() - interval 20 day);
-DELETE FROM casbotany.determination WHERE TimestampCreated  > (now() - interval 20 day);
-DELETE FROM casbotany.collectionobject WHERE TimestampCreated  > (now() - interval 20 day);
-DELETE FROM casbotany.collector WHERE TimestampCreated  > (now() - interval 20 day);
-DELETE FROM casbotany.collectingevent WHERE TimestampCreated  > (now() - interval 20 day);
-DELETE FROM casbotany.locality WHERE TimestampCreated  > (now() - interval 20 day);
-DELETE FROM casbotany.agent WHERE TimestampCreated  > (now() - interval 20 day);
+DELETE FROM collectionobjectattachment WHERE TimestampCreated  > (now() - interval 20 day);
+DELETE FROM attachment WHERE TimestampCreated  > (now() - interval 20 day);
+DELETE FROM determination WHERE TimestampCreated  > (now() - interval 20 day);
+DELETE FROM collectionobject WHERE TimestampCreated  > (now() - interval 20 day);
+DELETE FROM collector WHERE TimestampCreated  > (now() - interval 20 day);
+DELETE FROM collectingevent WHERE TimestampCreated  > (now() - interval 20 day);
+DELETE FROM locality WHERE TimestampCreated  > (now() - interval 20 day);
+DELETE FROM agent WHERE TimestampCreated  > (now() - interval 20 day);
 
 
 # Deletes leaf nodes from taxon tree: this may need to be run twice to remove multiple layers of new hierarchy
@@ -23,13 +23,13 @@ DELETE FROM casbotany.taxon WHERE TaxonID IN (SELECT TaxonID FROM temp_leaf_node
 
 DROP TEMPORARY TABLE IF EXISTS temp_leaf_nodes;
 
-ALTER USER 'botanist'@'%' ACCOUNT UNLOCK;
+ALTER USER 'user'@'%' ACCOUNT UNLOCK;
 
 
 # unlock command in case a script fails mid execution and fails to unlock database.
 UPDATE mysql.user
 SET account_locked = 'n'
-WHERE user != 'botanist' AND host = '%';
+WHERE user != 'user' AND host = '%';
 
 # commands used to kill multiple processes, used to deal with conflicts between open connections to sql db.
 #
